+++ examples/cloudflare-workers/worker/webpack.config.js
@@ -1,0 +1,9 @@
+module.exports = {
+  entry: "./index.js",
+  target: "webworker",
+  resolve: {
+    mainFields: ["main", "module"],
+    alias: {},
+  },
+  node: { fs: "empty" },
+};