+++ src/cli/bin.node.ts
@@ -6,11 +6,15 @@ const cli = cac('vite-plugin-ssr')
 cli
   .command('prerender')
   .option('--partial', 'allow only a subset of pages to be pre-rendered')
+  .option(
+    '--root <path>',
+    '[string] root directory of your project (where `vite.config.js` and `dist/` live) (default: `process.cwd()`)'
+  )
   .option('--client-router', 'serialize `pageProps` to JSON files for Client-side Routing')
   .option('--base <path>', `[string] public base path (default: /)`)
   .action(async (options) => {
-    const { partial, clientRouter, base } = options
-    await prerender({ partial, clientRouter, base })
+    const { partial, clientRouter, base, root } = options
+    await prerender({ partial, clientRouter, base, root })
   })
 
 // Listen to unknown commands
@@ -22,3 +26,7 @@ cli.help()
 cli.version(require('../../package.json').version)
 
 cli.parse(process.argv.length === 2 ? [...process.argv, '--help'] : process.argv)
+
+process.on('unhandledRejection', (rejectValue) => {
+  throw rejectValue
+})