+++ src/page-files/setup.client.ts
@@ -1,0 +1,3 @@
+import { setPageFiles } from './getPageFiles.shared'
+import { pageFiles } from './pageFiles.client'
+setPageFiles(pageFiles)