+++ src/plugin/utils.ts
@@ -1,0 +1,5 @@
+export { isSSR }
+
+function isSSR(config: { build?: { ssr?: boolean | string } }): boolean {
+  return !!config?.build?.ssr
+}