+++ src/utils/getFileUrl.ts
@@ -3,7 +3,7 @@ const pagePropsSuffix = '/index.pageProps.json'
 
 export { getFileUrl }
 export { isPagePropsUrl }
-export { retrieveOriginalUrl }
+export { removePagePropsSuffix }
 
 /**
  (`/`, `.html`) -> `/index.html`
@@ -18,19 +18,20 @@ function getFileUrl(url: string, fileExtension: '.html' | '.pageProps.json'): st
   assert(url.startsWith('/'))
   const { pathname, search, hash } = parseUrl(url)
   assert(url === `${pathname}${search}${hash}`)
-  const trailingHash = pathname.endsWith('/') ? '' : '/'
-  return `${pathname}${trailingHash}index${fileExtension}${search}${hash}`
+  const trailingSlash = pathname.endsWith('/') ? '' : '/'
+  return `${pathname}${trailingSlash}index${fileExtension}${search}${hash}`
 }
 
 function isPagePropsUrl(url: string): boolean {
   const { pathname } = parseUrl(url)
   return pathname.endsWith(pagePropsSuffix)
 }
-function retrieveOriginalUrl(url: string): string {
+function removePagePropsSuffix(url: string): string {
   assert(isPagePropsUrl(url))
-  let { pathname, search, hash } = parseUrl(url)
-  assert(url === `${pathname}${search}${hash}`)
+  let { origin, pathname, search, hash } = parseUrl(url)
+  assert(url === `${origin}${pathname}${search}${hash}`)
+  assert(pathname.endsWith(pagePropsSuffix))
   pathname = slice(pathname, 0, -1 * pagePropsSuffix.length)
   if (pathname === '') pathname = '/'
-  return `${pathname}${search}${hash}`
+  return `${origin}${pathname}${search}${hash}`
 }